+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## 0.0.2
+## 0.0.3
 
-* TODO:  new update release.
+### Added
+- `CustomLoader` widget.
+- `CustomText` and `CustomTextBold` widgets.
+- `CustomButton` widget.
+
+### Changed
+- Exported all widgets via `custom_feild_box.dart` for single import.